Move modern human readable functions out of modern pages
